combine cli commands into projects as flags
