+++ cmd_history.go
@@ -1,39 +1,11 @@
 package main
 
 import (
-	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
 	"time"
-
-	"github.com/spf13/cobra"
 )
-
-func newHistoryCmd() *cobra.Command {
-	var mode string
-	cmd := &cobra.Command{
-		Use:   "history",
-		Short: "List recently visited projects from history",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			items, err := readHistory()
-			if err != nil {
-				return err
-			}
-			if mode != "" {
-				for i, item := range items {
-					items[i] = mode + "\t" + item
-				}
-			}
-			if len(items) > 0 {
-				fmt.Print(strings.Join(items, "\n") + "\n")
-			}
-			return nil
-		},
-	}
-	cmd.Flags().StringVar(&mode, "mode", "", "Prefix each item with this mode (open|close) for use with fzf --with-nth")
-	return cmd
-}
 
 func parseTimestamp(s string) (time.Time, error) {
 	if t, err := time.Parse(time.RFC3339, s); err == nil {
